core: fix iio_register_get_ticks_us_cb never being defined
iio-backend.h declares iio_register_get_ticks_us_cb(), but library.c
defines iio_set_get_ticks_us_cb() instead, and defines it without __api.
The declared symbol has therefore never existed in libiio.so, and the
definition that does exist is hidden by -fvisibility=hidden. The feature
has been unreachable since it was introduced in be999f71, with no test
to notice.

Rename the definition to match the declaration so that the symbol is
actually exported, and add a test that registers a callback, emits a
timestamped log line, and checks the callback was called - and that
registering NULL goes back to the built-in clock.

Two further changes come with the rename:

The callback typedef returned unsigned int, but the value feeds
iio_read_counter_us(), which deals in a 64-bit microsecond count. On
platforms where unsigned int is 32-bit that wraps roughly every 72
minutes, and the truncation happens in the typedef, so no caller can
work around it. Widen it to uint64_t. This is an ABI change, but the
symbol was never callable, so nothing can depend on the old signature -
and after the 1.0 freeze it would no longer be fixable.

Re-base library_startup_time_us onto the new time source. Log timestamps
are reported relative to that value, which was sampled from the built-in
clock at load time; subtracting it from a reading of an unrelated clock
produces a meaningless delta.

+++ include/iio/iio-backend.h
@@ -45,9 +45,10 @@
 
 /*
  * Prototype of a user-supplied function that retrieves the number of ticks
- * (as microseconds) from a clock source.
+ * (as microseconds) from a clock source. The value is used as a 64-bit
+ * microsecond count; a 32-bit one would wrap after about 72 minutes.
  */
-typedef unsigned int (*iio_get_ticks_us)(void);
+typedef uint64_t (*iio_get_ticks_us)(void);
 
 struct iio_buffer;
 struct iio_device;
